Move extension methods to a new class and a new namespace (DSharpPlus.Interctivity.Extensions)
